requirements.txt
# Core
torch>=2.1
torchaudio>=2.1
hydra-core>=1.3
omegaconf>=2.3
transformers>=4.40
sentencepiece>=0.1.99
safetensors>=0.4
numpy>=1.22,<2.0
PyYAML>=6.0
huggingface_hub>=0.20
six>=1.16.0
tensorboard>=2.12
kaldialign>=0.7
lhotse>=1.20
tqdm>=4.65
soundfile>=0.12

